Move generic tests from RelationsGrpcClientsManagerTest to ChannelManagerTest.
