+++ src/main/java/org/project_kessel/clients/KesselClientsManager.java
@@ -8,18 +8,4 @@ public abstract class KesselClientsManager {
     protected KesselClientsManager(Channel channel) {
         this.channel = channel;
     }
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) return true;
-        if (o == null || getClass() != o.getClass()) return false;
-
-        KesselClientsManager that = (KesselClientsManager) o;
-        return channel.equals(that.channel);
-    }
-
-    @Override
-    public int hashCode() {
-        return channel.hashCode();
-    }
 }
